requirements.txt
pytest
pysqlite3-binary
chromadb
PyPDF2
openai
python-decouple
matplotlib
unidecode